+++ test/sgp4.jl
@@ -324,7 +324,7 @@ end
             "  │    RA of Asc. Node   : 162.1097°",
             "  │    Arg. of Periapsis : 136.2017°",
             "  │    Mean Anomaly      : 223.9283°",
-            "  │    B*                : 4.3e-5 1/er",
+            "  │    B*                : 4.3e-5 1/ER",
             "  ├─ Constants",
             "  │    R₀  : 6378.137 km",
             "  │    XKE : 0.07436685317 er^(3/2)/min",
@@ -341,6 +341,7 @@ end
     @test str == expected
 
     # The body can be printed under another header.
+    @test SatelliteToolboxSgp4.sgp4_is_initialized(sgp4d)
     str = sprint(SatelliteToolboxBase.print_tree_body, sgp4d)
     header = "Sgp4Propagator{Float64, Float64} (SGP4):\n"
     @test str == expected[(length(header) + 1):end]
@@ -370,6 +371,7 @@ end
 
     sgp4d = Sgp4Propagator{Float64}(SGP4C_WGS84)
 
+    @test !SatelliteToolboxSgp4.sgp4_is_initialized(sgp4d)
     @test repr(sgp4d) == "Sgp4Propagator{Float64, Float64} (not initialized)"
     expected = "Sgp4Propagator{Float64, Float64}:\n  Status : not initialized"
     @test sprint(show, MIME("text/plain"), sgp4d) == expected